+++ .storybook/main.js
@@ -4,15 +4,19 @@ module.exports = {
     "../libs/react/**/*.stories.mdx",
     "../libs/react/**/*.stories.@(js|jsx|ts|tsx)",
     "../src/**/*.stories.mdx",
-    "../src/**/*.stories.@(js|jsx|ts|tsx)"
   ],
   "addons": [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
     "@storybook/addon-a11y",
     "@storybook/addon-interactions",
-    // "@storybook/addon-storysource",
-    "@geometricpanda/storybook-addon-badges"
+    "@geometricpanda/storybook-addon-badges",
+    "@storybook/addon-storysource",
+    "@storybook/addon-coverage"
   ],
+  features: {
+    interactionsDebugger: true, // 👈 Enable playback controls
+    previewMdx2: true
+  },
   "framework": "@storybook/react"
 }