+++ fp-react/src/stories/Button.stories.tsx
@@ -1,40 +1,40 @@
-import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-
+import * as React from 'react';
+import { Story, Meta } from '@storybook/react';
 import { Button } from './Button';
+
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Example/Button',
   component: Button,
+  title: 'Example/Button',
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {
     backgroundColor: { control: 'color' },
   },
-} as ComponentMeta<typeof Button>;
+} as Meta;
 
-// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof Button> = (args) => <Button {...args} />;
 
-export const Primary = Template.bind({});
-// More on args: https://storybook.js.org/docs/react/writing-stories/args
+
+// ^?
+export const Primary: Story = ({args}) => <Button {...args} />;
 Primary.args = {
   primary: true,
   label: 'Button',
+  type: 'button',
 };
 
-export const Secondary = Template.bind({});
+export const Secondary: Story = ({args}) => <Button {...args} />;
 Secondary.args = {
   label: 'Button',
 };
 
-export const Large = Template.bind({});
+export const Large: Story = ({args}) => <Button {...args} />;
 Large.args = {
   size: 'large',
   label: 'Button',
 };
 
-export const Small = Template.bind({});
+export const Small: Story = ({args}) => <Button {...args} />;
 Small.args = {
   size: 'small',
   label: 'Button',