+++ libs/react/fp-badge/src/index.ts
@@ -1,0 +1,1 @@
+export { default as FP } from './component/fp';