+++ libs/react/fp-badge/vitest.config.js
@@ -1,0 +1,18 @@
+/// <reference types="vitest" />
+import { defineConfig } from 'vite'
+
+import react from '@vitejs/plugin-react'
+export default defineConfig({
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    setupFiles: './src/test/setup.ts',
+    coverage: {
+      reporter: ['text', 'html'],
+      exclude: [
+        'node_modules/',
+        'src/setupTests.ts',
+      ],
+    },
+  },
+});