+++ libs/react/fpkit/src/App.tsx
@@ -1,0 +1,21 @@
+import React, { useState } from 'react'
+import './App.css'
+
+import FPBadge from './component/badge/fp-badge'
+
+function App() {
+  const [count, setCount] = useState(0)
+
+  return (
+    <div className="App">
+      <header className="App-header">
+        <p>Hello Vite + React!</p>
+        <p>
+          <FPBadge elm="span" role='note'>Badge</FPBadge>
+        </p>
+      </header>
+    </div>
+  )
+}
+
+export default App