+++ libs/react/fpkit/src/index.ts
@@ -1,0 +1,6 @@
+import Badge from './component/badge/fp-badge'
+
+export { default as FP } from './component/fp'
+export {
+  Badge
+}